🔨 피드백 반영 : 프로젝트의 일관성 있는 구조를 위한 아키텍처 수정
