+++ src/main/java/kr/hhplus/be/server/application/concert/ConcertFacade.java
@@ -1,0 +1,15 @@
+package kr.hhplus.be.server.application.concert;
+
+import kr.hhplus.be.server.domain.concert.ConcertService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ConcertFacade {
+
+    private final ConcertService concertService;
+    private String test;
+
+
+}
