+++ src/main/java/kr/hhplus/be/server/application/user/UserFacade.java
@@ -1,0 +1,11 @@
+package kr.hhplus.be.server.application.user;
+
+import kr.hhplus.be.server.domain.user.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserFacade {
+    private final UserRepository userRepository;
+}
