+++ src/test/java/kr/hhplus/be/server/unit/ReservationFacadeTest.java
@@ -7,11 +7,10 @@ import kr.hhplus.be.server.domain.reservation.Reservation;
 import kr.hhplus.be.server.domain.reservation.ReservationService;
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.domain.user.UserService;
-import kr.hhplus.be.server.facade.reservation.ReservationFacade;
+import kr.hhplus.be.server.application.reservation.ReservationFacade;
 import kr.hhplus.be.server.presentation.reservation.dto.ReservationResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.BDDMockito;
 import org.mockito.InOrder;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
